Correct the commented issues.
Refactor the tests into separate files so they're easier to read.
